backend/requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854303569/work
bs4==0.0.1
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656783163977/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
conda==4.14.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1649385138064/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1657174084384/work
Flask==2.2.2
Flask-Cors==4.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.9
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1649385121707/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857351433/work
python-dotenv==1.0.1
pytz==2022.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1653464701544/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.2.post1
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
Werkzeug==2.2.2
